Mejoras Codigo Feedback
He hecho las mejoras que nos ha comentado Guille en los scripts de PlayerController, Bullet, Capture, AnimalController, CheckpointManager y UIManager.
Salvo el tema de flipear los sprites en vez de rotarlos sobre el eje Y, porque al hacerlo de esta manera no se rotaban los raycast de los animales y el firingPoint del jugador.

+++ Assets/Scripts/Animals/BarraDeSueño.cs
@@ -20,7 +20,7 @@ public class BarraDeSueño : MonoBehaviour
 {
     // ---- ATRIBUTOS DEL INSPECTOR ----
     #region Atributos del Inspector (serialized fields)
-    [SerializeField] int MaxBarraDeSueño;
+    [SerializeField] float MaxBarraDeSueño;
     [SerializeField] Collider2D TriggerCollider;
     [SerializeField] private Image _healthBarFill;
     [SerializeField] private float _fillSpeed;
@@ -57,7 +57,7 @@ public class BarraDeSueño : MonoBehaviour
     /// <summary>
     /// Método que se llama desde el script Bullet para aumentar el sueño del animal.
     /// </summary>
-    public void Dormir(int amount)
+    public void Dormir(float amount)
     {
         if(!dormido)
         {
